[refactor] FormLayout for repurposing layouts
